% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareGens.R
\name{CompareGens}
\alias{CompareGens}
\title{Compare the performance of two genotypes across all environments}
\usage{
CompareGens(GGEModel, G1, G2, ...)
}
\arguments{
\item{GGEModel}{An object of class "GGEModel" or "gge"}

\item{G1}{genotype to compare. Must be a string which perfectly matches a genotype label}

\item{G2}{genotype to compare. Must be a string which perfectly matches a genotype label and not equal to G1}

\item{...}{Other arguments sent to \code{\link[GGEBiplots]{GGEPlot}}}
}
\description{
Compare the performance of two genotypes across all environments
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=TRUE)
GGE<-GGEModel(GxEMeans)
CompareGens(GGE,"Unica","Canchan")
}
\keyword{GGE}
