% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscRep.R
\name{DiscRep}
\alias{DiscRep}
\title{Discrimination vs. representativeness}
\usage{
DiscRep(GGEModel, ...)
}
\arguments{
\item{GGEModel}{An object of class "GGEModel" or "gge"}

\item{...}{Other arguments sent to \code{\link[GGEBiplots]{GGEPlot}}}
}
\description{
Evaluating the environments based on both discriminating ability and representativeness
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=TRUE)
GGE<-GGEModel(GxEMeans)
DiscRep(GGE)
}
\keyword{GGE}
