% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvRelationship.R
\name{EnvRelationship}
\alias{EnvRelationship}
\title{Relationship between environments}
\usage{
EnvRelationship(GGEModel, ...)
}
\arguments{
\item{GGEModel}{A GGEModel produced from a call to GGEModels()}

\item{...}{Other arguments sent to GGEPlot()}
}
\description{
Relationship between environments
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=T)
GGE<-GGEModels(GxEMeans)
GGEPlot(GGE)
EnvRelationship(GGE)
}
\keyword{GGE}
