% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExamineEnv.R
\name{ExamineEnv}
\alias{ExamineEnv}
\title{Examine an environment}
\usage{
ExamineEnv(GGEModel, Env, ...)
}
\arguments{
\item{GGEModel}{An object of class "GGEModel" or "gge"}

\item{Env}{environment to examine. Must be a string which perfectly matches an environment label}

\item{...}{Other arguments sent to \code{\link[GGEBiplots]{GGEPlot}}}
}
\description{
Ranking the cultivars based on their performance in any given environment
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=TRUE)
GGE<-GGEModel(GxEMeans)
ExamineEnv(GGE,"Ayac")
}
\keyword{GGE}
