% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExamineGen.R
\name{ExamineGen}
\alias{ExamineGen}
\title{Examine a genotype}
\usage{
ExamineGen(GGEModel, selectedE, ...)
}
\arguments{
\item{GGEModel}{A GGEModel produced from a call to GGEModels()}

\item{...}{Other arguments sent to GGEPlot()}

\item{selectedG}{genotype to examine. Must be a string which perfectly matches an genotype label}
}
\description{
Examine a genotype
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=T)
GGE<-GGEModels(GxEMeans)
GGEPlot(GGE)
ExamineGen(GGE,selectedG="Unica")
}
\keyword{GGE}
