% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGEModel.R
\name{GGEModel}
\alias{GGEModel}
\title{Produces GGE model from 2-way table of means}
\usage{
GGEModel(Data, centering = "tester", scaling = "none", SVD = "column")
}
\arguments{
\item{Data}{wide data frame containing GxE means with genotypes in rows and environments in columns with rownames and colnames indicating genotype names and environment names}

\item{centering}{centering method. Options are "tester", "global","double" or "none" for tester centered (G+GE), global centered (E+G+GE), double centred (GE) or no centering. 
Models produced without centering cannot be used in the GGEPlot() function. Defaults to tester centered.}

\item{scaling}{scaling method. Options are "sd" or "none"}

\item{SVD}{method for singular value partitioning. Options are "row","column","dual" or "symmetrical". Defaults to "column"}
}
\description{
Calculates the GGE model where presented with a two way table of means with genotypes in rows (and genotype names as row names) and environments in columns (and environment names as column names)
Code to produce the model taken with very few modifications from internal code within \code{\link[GGEBiplotGUI]{GGEBiplot}}
For dealing with missing data then please use the \code{\link[gge]{gge}}
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=TRUE)
GGE<-GGEModel(GxEMeans)
GGEPlot(GGE)
}
\keyword{GGE}
