% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGEModels.R
\name{GGEModels}
\alias{GGEModels}
\title{GGE Models for Input into GGE Biplots and for summarising explained variability per component}
\usage{
GGEModels(Data, centering = "tester", scaling = "none", SVP = "column")
}
\arguments{
\item{Data}{wide data frame containing GxE means with genotypes in rows and environments in columns with rownames and colnames indicating genotype names and environment names}

\item{centering}{centering method. Options are "tester", "global","double" or "none" for tester centered (G+GE), global centered (E+G+GE), double centred (GE) or no centering. 
No centering is strongly not reccomended as some functionality will not work. Defaults to tester centered.}

\item{scaling}{scaling method. Options are "sd" or "none"}

\item{SVP}{method for singular value partitioning. Options are "row","column","dual" or "symmetrical". Defaults to "column"}
}
\description{
GGE Models for Input into GGE Biplots and for summarising explained variability per component
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=T)
GGEModels(GxEMeans)
}
\keyword{GGE}
