% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGEPlot.R
\name{GGEPlot}
\alias{GGEPlot}
\title{Main plotting function for GGEBiplots. Produces basic biplot unless otherwise specified. 
Most attributes of graph can either be customised within the function (text size and color) or turned off and then customised by the user using standard ggplot2 commands (titles and axes)}
\usage{
GGEPlot(GGEModel, type = 1, d1 = 1, d2 = 2, selectedE = NA,
  selectedG = NA, selectedG1 = NA, selectedG2 = NA,
  colGen = "forestgreen", colEnv = "blue", sizeGen = 4, sizeEnv = 4,
  largeSize = 4.5, axis_expand = 1.4, axislabels = TRUE, axes = TRUE,
  limits = TRUE, titles = TRUE, footnote = TRUE)
}
\arguments{
\item{GGEModel}{A GGEModel produced from a call to GGEModels()}

\item{type}{type of biplot to produce. 1=Basic biplot; 2=Examine environment; 3=Examine genotype; 4=Relationship among environments; 5=Compare two genotypes;
6=Which won where/what; 7=Discrimination vs. representativeness; 8=Relationship among environments; 9=Mean vs. stability; 10=Relationship among genotypes.
Defaults to type=1. All other types have available wrapper functions}

\item{d1}{component to plot on x axis. Defaults to 1}

\item{d2}{component to plot on y axis. Defaults to 1}

\item{selectedE}{environment to examine when type=2. Must be a string which perfectly matches an environment label}

\item{selectedG}{genotype to examine when type=3. Must be a string which perfectly matches a genotype label}

\item{selectedG1}{genotype to compare when type=5. Must be a string which perfectly matches a genotype label}

\item{selectedG2}{genotype to compare when type=5. Must be a string which perfectly matches a genotype label and not equal to selectedG1}

\item{colGen}{colour for genotype attributes on biplot. Defaults to "forestgreen"}

\item{colEnv}{colour for environment attributes on biplot. Defaults to "blue"}

\item{sizeGen}{text size for genotype labels on biplot. Defaults to 4}

\item{sizeEnv}{text size for environment labels on biplot. Defaults to 4}

\item{largeSize}{text size to use for larger labels on type=5 (for the two selected genotypes) and type=6 (for the "winners"). Defaults to 4.5}

\item{axis_expand}{multiplication factor to expand the axis limits by to enable fitting of labels. Defaults to 1.4}

\item{axislabels}{TRUE/FALSE. Include axis labels}

\item{axes}{TRUE/FALSE. Include axes modifications}

\item{limits}{TRUE/FALSE. Include calls to scale_x_continuous() and scale_y_continuous()}

\item{titles}{TRUE/FALSE. Include title}

\item{footnote}{TRUE/FALSE. Include footnote}
}
\description{
Main plotting function for GGEBiplots. Produces basic biplot unless otherwise specified. 
Most attributes of graph can either be customised within the function (text size and color) or turned off and then customised by the user using standard ggplot2 commands (titles and axes)
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=T)
GGE<-GGEModels(GxEMeans)
GGEPlot(GGE)
}
\keyword{GGE}
