% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanStability.R
\name{MeanStability}
\alias{MeanStability}
\title{Mean vs Stability}
\usage{
MeanStability(GGEModel, ...)
}
\arguments{
\item{GGEModel}{A GGEModel produced from a call to GGEModels()}

\item{...}{Other arguments sent to GGEPlot()}
}
\description{
Mean vs Stability
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=T)
GGE<-GGEModels(GxEMeans)
GGEPlot(GGE)
MeanStability(GGE)
}
\keyword{GGE}
