\name{Ontario}
\alias{Ontario}
\docType{data}
\title{
Ontario winter wheat (1993)
}
\description{
The sample data are yields from the 1993 Ontario winter wheat (Triticum aestivum L.) performance
trials, in which 18 cultivars were tested at nine locations (Yan and Kang 2003).
Duplicated from the archived package 'GGEBiplotGUI'.
}
\usage{data(Ontario)}
\format{
  A data frame with 18 observations on the following 10 variables.
}
\source{
Yan W, Kang MS (2003). \dQuote{GGE Biplot Analysis: A Graphical Tool for Breeders, Geneticists,
and Agronomists.} \emph{CRC Press, Boca Raton, FL, USA}.
}
\examples{
data(Ontario)
}
\keyword{datasets}
