% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankEnv.R
\name{RankEnv}
\alias{RankEnv}
\title{Ranking environments}
\usage{
RankEnv(GGEModel, ...)
}
\arguments{
\item{GGEModel}{A GGEModel produced from a call to GGEModels()}

\item{...}{Other arguments sent to GGEPlot()}
}
\description{
Ranking environments
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=T)
GGE<-GGEModels(GxEMeans)
GGEPlot(GGE)
RankEnv(GGE)
}
\keyword{GGE}
