% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankGen.R
\name{RankGen}
\alias{RankGen}
\title{Ranking Genotypes}
\usage{
RankGen(GGEModel, ...)
}
\arguments{
\item{GGEModel}{A GGEModel produced from a call to GGEModels()}

\item{...}{Other arguments sent to GGEPlot()}
}
\description{
Ranking Genotypes
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=T)
GGE<-GGEModels(GxEMeans)
GGEPlot(GGE)
RankGen(GGE)
}
\keyword{GGE}
