% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stattable.R
\name{stattable}
\alias{stattable}
\title{Produce a 2 way summary table of results}
\usage{
stattable(rowfactor, columnfactor, outcome, FUN = mean, ...)
}
\arguments{
\item{rowfactor}{vector containing factor variable to be included in the rows}

\item{columnfactor}{vector containing facctor variable to be included in the rows}

\item{outcome}{vector containing numeric outcome values}

\item{FUN}{name of summary function to use}

\item{...}{other arguments for FUN}
}
\description{
Exists as a convenience function to transform raw data into a nice format for
other functions by design rather than as a side-effect of combining list with
tapply
}
\examples{
fakedata<-data.frame(expand.grid(Genotype=1:10,Environment=1:10,Rep=1:3),Outcome=rnorm(300))
stattable(fakedata$Genotype,fakedata$Environment,fakedata$Outcome,na.rm=TRUE)
}
\keyword{2way}
\keyword{GGEBiplotGUI}
\keyword{means}
\keyword{statistics}
\keyword{summary}
\keyword{table}
