% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhichWon.R
\name{WhichWon}
\alias{WhichWon}
\title{Which won where/what}
\usage{
WhichWon(GGEModel, axis_expand = 1.6, ...)
}
\arguments{
\item{GGEModel}{An object of class "GGEModel" or "gge"}

\item{axis_expand}{multiplication factor to expand the axis limits by to enable fitting of labels. Defaults to 1.6 for the Which Won Plot}

\item{...}{Other arguments sent to \code{\link[GGEBiplots]{GGEPlot}}}
}
\description{
Identifying the 'best' cultivar in each environment
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=TRUE)
GGE<-GGEModel(GxEMeans)
WhichWon(GGE)
}
\keyword{GGE}
