% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareGens.R
\name{CompareGens}
\alias{CompareGens}
\title{Compare two genotypes}
\usage{
CompareGens(GGEModel, selectedG1, selectedG2, ...)
}
\arguments{
\item{GGEModel}{A GGEModel produced from a call to GGEModels()}

\item{selectedG1}{genotype to compare when type=5. Must be a string which perfectly matches a genotype label}

\item{selectedG2}{genotype to compare when type=5. Must be a string which perfectly matches a genotype label and not equal to selectedG1}

\item{...}{Other arguments sent to GGEPlot()}
}
\description{
Compare two genotypes
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=T)
GGE<-GGEModels(GxEMeans)
GGEPlot(GGE)
CompareGens(GGE,selectedG1="Unica",selectedG2="Canchan")
}
\keyword{GGE}
