% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareGens.R
\name{CompareGens}
\alias{CompareGens}
\title{Compare two genotypes biplot}
\usage{
CompareGens(GGEModel, G1, G2, ...)
}
\arguments{
\item{GGEModel}{An object of class \code{GGEModel} or \code{gge}}

\item{G1}{genotype to compare. Must be a string which matches a genotype label}

\item{G2}{genotype to compare. Must be a string which matches a genotype label and not equal to G1}

\item{...}{Other arguments sent to \code{\link[GGEBiplots]{GGEPlot}}}
}
\description{
Compare the performance of two genotypes across all environments
}
\examples{
library(GGEBiplotGUI)
data(Ontario)
GGE1<-GGEModel(Ontario)
CompareGens(GGE1,"cas","luc")
}
\keyword{GGE}
