% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscRep.R
\name{DiscRep}
\alias{DiscRep}
\title{Discrimination vs. representativeness}
\usage{
DiscRep(GGEModel, ...)
}
\arguments{
\item{GGEModel}{A GGEModel produced from a call to GGEModels()}

\item{...}{Other arguments sent to GGEPlot()}
}
\description{
Discrimination vs. representativeness
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=T)
GGE<-GGEModels(GxEMeans)
GGEPlot(GGE)
DiscRep(GGE)
}
\keyword{GGE}
