% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscRep.R
\name{DiscRep}
\alias{DiscRep}
\title{Discrimination vs. representativeness biplot}
\usage{
DiscRep(GGEModel, ...)
}
\arguments{
\item{GGEModel}{An object of class \code{GGEModel} or \code{gge}}

\item{...}{Other arguments sent to \code{\link[GGEBiplots]{GGEPlot}}}
}
\description{
Evaluating the environments based on both discriminating ability and representativeness
}
\examples{
data(Ontario)
GGE1<-GGEModel(Ontario)
DiscRep(GGE1)
}
\keyword{GGE}
