% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvRelationship.R
\name{EnvRelationship}
\alias{EnvRelationship}
\title{Relationship between environments}
\usage{
EnvRelationship(GGEModel, ...)
}
\arguments{
\item{GGEModel}{An object of class "GGEModel" or "gge"}

\item{...}{Other arguments sent to \code{\link[GGEBiplots]{GGEPlot}}}
}
\description{
Relationship between environments
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=TRUE)
GGE<-GGEModel(GxEMeans)
EnvRelationship(GGE)
}
\keyword{GGE}
