% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvRelationship.R
\name{EnvRelationship}
\alias{EnvRelationship}
\title{Relationship between environments}
\usage{
EnvRelationship(GGEModel, ...)
}
\arguments{
\item{GGEModel}{An object of class \code{GGEModel} or \code{gge}}

\item{...}{Other arguments sent to \code{\link[GGEBiplots]{GGEPlot}}}
}
\description{
Relationship between environments
}
\examples{
data(Ontario)
GGE1<-GGEModel(Ontario)
EnvRelationship(GGE1)
}
\keyword{GGE}
