% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExamineEnv.R
\name{ExamineEnv}
\alias{ExamineEnv}
\title{Examine an environment}
\usage{
ExamineEnv(GGEModel, selectedE, ...)
}
\arguments{
\item{GGEModel}{A GGEModel produced from a call to GGEModels()}

\item{selectedE}{environment to examine. Must be a string which perfectly matches an environment label}

\item{...}{Other arguments sent to GGEPlot()}
}
\description{
Examine an environment
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=T)
GGE<-GGEModels(GxEMeans)
GGEPlot(GGE)
ExamineEnv(GGE,selectedE="Ayac")
}
\keyword{GGE}
