% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExamineEnv.R
\name{ExamineEnv}
\alias{ExamineEnv}
\title{Examine an environment}
\usage{
ExamineEnv(GGEModel, Env, ...)
}
\arguments{
\item{GGEModel}{An object of class \code{GGEModel} or \code{gge}}

\item{Env}{environment to examine. Must be a string which matches an
environment label}

\item{...}{Other arguments sent to \code{\link[GGEBiplots]{GGEPlot}}}
}
\description{
Ranking the cultivars based on their performance in any given environment
}
\examples{
library(GGEBiplotGUI)
data(Ontario)
GGE1<-GGEModel(Ontario)
ExamineEnv(GGE1,"WP93")
}
\keyword{GGE}
