% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExamineGen.R
\name{ExamineGen}
\alias{ExamineGen}
\title{Examine a genotype}
\usage{
ExamineGen(GGEModel, Gen, ...)
}
\arguments{
\item{GGEModel}{An object of class "GGEModel" or "gge"}

\item{Gen}{genotype to examine. Must be a string which perfectly matches an genotype label}

\item{...}{Other arguments sent to \code{\link[GGEBiplots]{GGEPlot}}}
}
\description{
Ranking the environments based on the relative performance of any given cultivar
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=TRUE)
GGE<-GGEModel(GxEMeans)
ExamineGen(GGE,"Unica")
}
\keyword{GGE}
