% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExamineGen.R
\name{ExamineGen}
\alias{ExamineGen}
\title{Examine a genotype biplot}
\usage{
ExamineGen(GGEModel, Gen, ...)
}
\arguments{
\item{GGEModel}{An object of class \code{GGEModel} or \code{gge}}

\item{Gen}{genotype to examine. Must be a string which perfectly matches an genotype label}

\item{...}{Other arguments sent to \code{\link[GGEBiplots]{GGEPlot}}}
}
\description{
Ranking the environments based on the relative performance of any given cultivar
}
\examples{
library(GGEBiplotGUI)
data(Ontario)
GGE1<-GGEModel(Ontario)
ExamineGen(GGE1,"cas")
}
\keyword{GGE}
