% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanStability.R
\name{MeanStability}
\alias{MeanStability}
\title{Mean vs Stability}
\usage{
MeanStability(GGEModel, ...)
}
\arguments{
\item{GGEModel}{An object of class "GGEModel" or "gge"}

\item{...}{Other arguments sent to \code{\link[GGEBiplots]{GGEPlot}}}
}
\description{
Evaluating cultivars based on both average yield and stability
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=TRUE)
GGE<-GGEModel(GxEMeans)
MeanStability(GGE)
}
\keyword{GGE}
