% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanStability.R
\name{MeanStability}
\alias{MeanStability}
\title{Mean vs. Stability Biplot}
\usage{
MeanStability(GGEModel, ...)
}
\arguments{
\item{GGEModel}{An object of class \code{GGEModel} or \code{gge}}

\item{...}{Other arguments sent to \code{\link[GGEBiplots]{GGEPlot}}}
}
\description{
Evaluating cultivars based on both average yield and stability
}
\examples{
data(Ontario)
GGE1<-GGEModel(Ontario)
MeanStability(GGE1)
}
\keyword{GGE}
