% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankGen.R
\name{RankGen}
\alias{RankGen}
\title{Ranking genotypes with respect to the ideal genotype}
\usage{
RankGen(GGEModel, ...)
}
\arguments{
\item{GGEModel}{An object of class "GGEModel" or "gge"}

\item{...}{Other arguments sent to \code{\link[GGEBiplots]{GGEPlot}}}
}
\description{
Ranking genotypes with respect to the ideal genotype
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=TRUE)
GGE<-GGEModel(GxEMeans)
RankGen(GGE)
}
\keyword{GGE}
