% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankGen.R
\name{RankGen}
\alias{RankGen}
\title{Ranking genotypes with respect to the ideal genotype}
\usage{
RankGen(GGEModel, ...)
}
\arguments{
\item{GGEModel}{An object of class \code{GGEModel} or \code{gge}}

\item{...}{Other arguments sent to \code{\link[GGEBiplots]{GGEPlot}}}
}
\description{
Ranking genotypes with respect to the ideal genotype
}
\examples{
library(GGEBiplotGUI)
data(Ontario)
GGE1<-GGEModel(Ontario)
RankGen(GGE1)
}
\keyword{GGE}
