% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhichWon.R
\name{WhichWon}
\alias{WhichWon}
\title{Which won where/what}
\usage{
WhichWon(GGEModel, axis_expand = 1.6, ...)
}
\arguments{
\item{GGEModel}{A GGEModel produced from a call to GGEModels()}

\item{...}{Other arguments sent to GGEPlot()}
}
\description{
Which won where/what
}
\examples{
library(agricolae)
data(plrv)
GxEMeans<-tapply(plrv$Yield,list(plrv$Genotype,plrv$Locality),mean,na.rm=T)
GGE<-GGEModels(GxEMeans)
GGEPlot(GGE)
WhichWon(GGE)
}
\keyword{GGE}
