% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhichWon.R
\name{WhichWon}
\alias{WhichWon}
\title{Which Won Where/What Biplot}
\usage{
WhichWon(GGEModel, ...)
}
\arguments{
\item{GGEModel}{An object of class \code{GGEModel} or \code{gge}}

\item{...}{Other arguments sent to \code{\link[GGEBiplots]{GGEPlot}}}
}
\description{
Identifying the 'best' cultivar in each environment
}
\examples{
data(Ontario)
GGE1<-GGEModel(Ontario)
WhichWon(GGE1)
}
\keyword{GGE}
